Refactor into ES6 modules
